Cleanup: use new forge onStopUsing hook to improve finish using modifier hooks
Removes the need for our equipment change watcher to clear that mostly, and removes some possibly redundant clears.
Also sets us up for a onStopUsing modifier hook, not sure if we want that or not yet. Couldn't before because the modifier would be cleared too early but we can guarantee onStopUsing is the last using hook to be called