Migrate to new Mantle tag conditions
The most notable change is Mantle's new `tag_filled` condition, which is a much more compact version of the common `not(tag_empty)` pattern. Still no idea why Forge decided to make the condition for an empty tag when filled is far more useful.
In addition, both the difference and intersection tag predicates are replaced with the new tag combination predicate. This predicate matches if the intersection of the listed tags ignoring the listed tag is non-empty, making it much more flexible than either variant. The one downside is you can't do tag OR with it, as that was redundant to just making a combined tag (which we now do with ore rates)