Allow opening a tool inventory menu by right clicking the tool from the inventory
Also works with any whitelisted containers, like the tool inventory
Not supported in non-inventory menus as there is no way to return to your previous container. Though I want to see if such a thing is possible.
As part of this, the tool inventory screen now has a slot index instead of an equipment slot tracking the slot in charge.
Also discovered a related bug that I'll want to backport, the offhand and mainhand indexes are computed wrongly for one of the methods in the container