Fix Json Things tools not supporting stack sizes larger than 1
Since ModifiableItem extends TieredItem, it was forcing the stack size to 1. Fix by just overriding the stack size getter and adding it as a constructor parameter
Has the added benefit of tool items added via Json Things now default to stack size 1
Daggers use both the old hack and the new fix as it gives better reporting if someone calls the contextless getter.