More item use interactions
In creative mode, can now use creative slots and modifier crystals to add the appropriate slot/modifier for free
In survival mode, modifier crystals remain unusable for their previous functions, costing a slot when used in the tinkers anvil properly
Repair kits work in both modes to repair a tool on right click, saves effort using the crafting window

Due to these changes, we no longer need the creative slot recipe, so it has been removed. If you are a modpack maker wishing to give your players creative slots, don't, that is not what they are for. We have a ton of flexability adding new ways for your players to get modifiers instead of crafting a thing not designed for survival.