Fix too large smelteries having too many slots, breaking inventory sync
Turns out we can 10920 slots before syncing breaks due to the indexes syncing as shorts. So, cap the number of slots to that number to prevent issues.
Means we deprive people of their slots in their giant smeltery, but eh, they won't notice. Its not like you can actually use a smeltery that big without lagging your game to a crawl.