Fix melter not detecting heaters (#5356)
Worth noting that while this does fix the bug, this fix is incomplete.
The melter actually has two bugs right now - the heater is never checked (as the misplaced else which is fixed by this commit), and more importantly once it decides it has no tank, it never will detect a tank.
Former was an easy fix, for the latter I think I need to separate the modules as the caches for multi-tank is getting in the way of the desired usage in single tank; melter doesn't have a good spot to clear the cache without a hack that would be wrong for smeltery/foundry