Split fuel module into two pieces since the smeltery and melter are so different
SolidFuelModule handles melter and alloyer. This notably does away with a lot of caches that are just excessive for the melters single tank design. This fixes a bug with said cache not being cleared properly when the melter changes.
MultitankFuelModule handles smeltery and foundry. This contains all the advanced caching logic that is used by the smeltery. Some cleanup might have made the fuel change more responsive.

Minor bugs fixed while making this:
* Fixed wrong size for container listeners in multitank fuel module
* Fix wrong tank being used in melter menu interactions