Restrict survival operator usage of creative slots and modifier crystals to a config
To be clear, anyone can continue to use these in creative.
Before, survival players could use it if they are operators (i.e. have cheats enabled). Now, survival players must be both an operator and have enabled the config option to use it.
This should minimize confusion with using the feature for modpack devs (see #5358), as they are unlikely to use the feature for testing unless they are adding new tinkers content. But keeps it available for me who finds it super useful in dev testing.