Fix crafting table not calling onCraftedBy on the wrong stack instance
Sideeffects notably include maps not getting properly initialized when scaling up (#5360)
The first part of the fix is to ensure that we propagate the taken stack in `onTake` to the inventory instead of recalculating it.
To ensure that is the correct stack, we could override `tryRemove`, but since the crafting station already has player access via the slot in `PlayerSensitiveLazyResultSlot`, just make that more stable
To be specific, `PlayerSensitiveLazyResultSlot` now caches the per player result directly instead of using the block entity cache. This ensures the first player to look at the result doesn't determine the recipe.
Non-crafting station block entities now use methods that make it more clear the player is ignored for their crafting through the menus; we don't want player sensitive modifier recipes at this time.