Improve ichor and ender fluid effects
Ichor's "break any block" was a bit redundant to enderslime, which does the same provided the offhand is not a block.
Place block on ender was a little expensive for practical use, so giving that to ichor seemed fitting.
Ender now does a bit more powerful interact at range, very useful on blocks and moderately useful on entities. Still can technically place blocks but may not prioritize that. It might be a little janky as currently I do no client side interaction, just serverside; will have to see if that causes issues.

I do want to allow the fluid cannon to use both ichor and maybe ender, though likely need a fake player for that. Placing blocks at least might be able to get away without one