Implement the proxy tank
The proxy tank is a block that redirects fluid calls to the contained item. Its designed for filling and emptying the tanks of tools.
The proxy tank by default blacklists any fluid handler items with a minimum unit larger than 1 (such as buckets and copper cans) as that leads to some weird behavior. Debatably there is a faucet bug that needs fixing here, specifically with tanks of exactly 90mb.

For buckets, copper cans, and mantle fluid transfer (e.g. soup bowls), the better block to use is the casting tank, see #5381 for current in progress implementation.